add string format for text values
Lua's 5.3 release adds integers and caused some other notable changes
around the behaviour of floats being cast to strings [1]. This causes
certain widgets to report "silly" values for certain resources like 33.0%
RAM usage instead of just 33%.

Note, this is likely not the best way to fix this problem. A better
solution may be to have the user specify a `text_format` property and
default, maybe, to `%2.f` if nothing is provided. Will explore.

[1]: http://www.lua.org/manual/5.3/manual.html#